Minor fixes
