+++ rmom-manager/sqlx-data.json
@@ -248,26 +248,6 @@
     },
     "query": "SELECT q.label, m.host, m.port FROM queue q INNER JOIN mom m ON q.mom_id = m.id WHERE q.label = $1"
   },
-  "a1d32b81e81956c970fef07c5f210de6ec2bbf176939e9e068aa9236e06efac1": {
-    "describe": {
-      "columns": [
-        {
-          "name": "topic",
-          "ordinal": 0,
-          "type_info": "Varchar"
-        }
-      ],
-      "nullable": [
-        false
-      ],
-      "parameters": {
-        "Left": [
-          "Text"
-        ]
-      }
-    },
-    "query": "SELECT DISTINCT channel.topic FROM channel INNER JOIN queue ON channel.queue_id = queue.id WHERE queue.label = $1 AND channel.topic != '__none__'"
-  },
   "a42074132f37f8b3fb6da1d074d88dfb7433433fe26c05ac9024fc179a0a9da2": {
     "describe": {
       "columns": [
